v0.9.3: incendiary darts now use durability when used on terrain

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/weapon/missiles/darts/IncendiaryDart.java
@@ -42,7 +42,12 @@ public class IncendiaryDart extends TippedDart {
 		Char enemy = Actor.findChar( cell );
 		if ((enemy == null || enemy == curUser) && Dungeon.level.flamable[cell]) {
 			GameScene.add(Blob.seed(cell, 4, Fire.class));
-			Dungeon.level.drop(new Dart(), cell).sprite.drop();
+			decrementDurability();
+			if (durability > 0){
+				super.onThrow(cell);
+			} else {
+				Dungeon.level.drop(new Dart(), cell).sprite.drop();
+			}
 		} else{
 			super.onThrow(cell);
 		}
